Add rules for list fusion
Make `fromList` fuse with `build` and `augment`, and make `toList` fuse
with `foldr`. Express the `toList/fromList` rule in terms of rewrite
helper targets; this has the pleasant side effect of letting it handle
things like `fromList $ toList xs ++ toList ys`.

Adjust arities of functions in `Data.BitSet.Word` to convince
most of the underlying functions to inline. I *think* this is
what we want.